Move components into subfolder
