+++ lib/responseDownloadFormats/html/components/ProtectedWarning.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { SecurityAttribute } from "@lib/types";
-import { customTranslate } from "../helpers";
+import { customTranslate } from "../../helpers";
 
 // Note: use lang prop if you want to force a specific language e.g. HTMLDownload component
 interface ProtectedWarningProps {
